Rename match_to_purld to match_to_matchcode
    * Move matchcode related code from pipes/purldb.py to pipes/matchcode.py
    * Update scancode.io settings to get matchcode API settings
    * This is done since matchcode and purldb are technically two seperate services with their own users, etc.
