+++ setup.cfg
@@ -136,7 +136,7 @@ scancodeio_pipelines =
     load_inventory = scanpipe.pipelines.load_inventory:LoadInventory
     load_sbom = scanpipe.pipelines.load_sbom:LoadSBOM
     map_deploy_to_develop = scanpipe.pipelines.deploy_to_develop:DeployToDevelop
-    match_to_purldb = scanpipe.pipelines.match_to_purldb:MatchToPurlDB
+    match_to_matchcode = scanpipe.pipelines.match_to_matchcode:MatchToMatchCode
     populate_purldb = scanpipe.pipelines.populate_purldb:PopulatePurlDB
     resolve_dependencies = scanpipe.pipelines.resolve_dependencies:ResolveDependencies
     scan_codebase = scanpipe.pipelines.scan_codebase:ScanCodebase
